+++ admin-frontend/src/router/guards/check-login.ts
@@ -1,4 +1,4 @@
-import type { Router } from 'vue-router'
+import type { NavigationGuardReturn, Router } from 'vue-router'
 import { redirectAfterLoginBeforeRoute, redirectToChooseIdentityBeforeRoute, redirectToLoginBeforeRoute } from '@/router/guards/scripts/redirect_to'
 import { useIdentityStore } from '@/stores/system/identity'
 import { useUserStore } from '@/stores/user'
@@ -7,7 +7,7 @@ import { useUserStore } from '@/stores/user'
  * 用户登录校验 路由守卫
  */
 export function createCheckLoginGuard(router: Router) {
-  router.beforeEach(async (to, from, next) => { // router.currentRoute.value === from
+  router.beforeEach(async (to, from): Promise<NavigationGuardReturn> => { // router.currentRoute.value === from
 
     const userStore = useUserStore()
     const identityStore = useIdentityStore()
@@ -26,8 +26,7 @@ export function createCheckLoginGuard(router: Router) {
         // 已登录, 访问登录页时跳转
         // console.log('已登录, 在登录页')
         // 如果有 redirect_url 则跳转，否则跳到 dashboard
-        redirectAfterLoginBeforeRoute(to, next)
-        return
+        return redirectAfterLoginBeforeRoute(to)
       } else if (isInChooseIdentityPage) {
         // 已登录, 在切换身份页
         // console.log('已登录, 在切换身份页')
@@ -37,8 +36,7 @@ export function createCheckLoginGuard(router: Router) {
         // console.log('已登录, 不在登录页、切换身份页')
         if (!identityStore.currentIdentity) {
           if (identityStore.userIdentityDtoList && identityStore.userIdentityDtoList.length > 1) {
-            redirectToChooseIdentityBeforeRoute(to, next)
-            return
+            return redirectToChooseIdentityBeforeRoute(to)
           }
         }
       }
@@ -49,10 +47,9 @@ export function createCheckLoginGuard(router: Router) {
       } else {
         // 未登录, 不在登录页时跳转
         // console.log('未登录, 不在登录页')
-        redirectToLoginBeforeRoute(from, next)
-        return
+        return redirectToLoginBeforeRoute(from)
       }
     }
-    next()
+
   })
 }